Specify and use dependencies instead of depending on globals.
Only core dependencies are specified. Other dependencies will be specified
under #88.

The core dependencies are Backbone, jQuery, and underscore. All of these
are specified as peerDependencies, requiring the consumer to ensure they're
satisfied.

Backbone and underscore are specified as devDependencies to be used in specs.
jQuery is provided to specs via a vendor/ file since the current version we're
using isn't available as a node module.